Move irc to pkg/irc
