+++ pkg/irc/parser.go
@@ -1,4 +1,4 @@
-package parser
+package irc
 
 import (
 	"strings"
