+++ pkg/irc/parser_test.go
@@ -1,4 +1,4 @@
-package parser
+package irc
 
 import "testing"
